Extracted Total of Cases and Deaths Per Continent
This excel file contains extracted to number of death and number of cases that were obtained from each continent due to covid virus infections. These are grouped into continents over a given period of time. The contains a graphical representation of how total number of death is highly correlated with the total number of cases in each continent.