Uganda Covid Info
This is a Ugandan excel covid file that shows extracted data from the original raw file. This shows total cases, new cases and deaths daily for a given period of time. Rows with that had null data were replaced with 0 as a standard value using numpy and pandas libraries that are used in python programming language.